Remove the remaining python attach calls from reader, writer and params
