+++ crates/qpy/src/circuit_writer.rs
@@ -211,18 +211,22 @@ fn pack_instruction_blocks(
         Parameters::Params(_) => Err(QpyError::ConversionError(
             "Instruction has params but expected blocks".to_string(),
         )),
-        Parameters::Blocks(blocks) => Python::attach(|py| -> Result<_, QpyError> {
-            blocks
-                .iter()
-                .map(|block: &CircuitData| -> Result<_, QpyError> {
-                    // we explicitly name the block "unnamed" because otherwise it will be assigned a serial number name (e.g. "circuit-45")
-                    // which would result in inconsistent results, e.g. when packing the same circuit twice on the same run
-                    let py_block: PyCircuitData = block.clone().into();
-                    let circuit = py_block.into_py_quantum_circuit(py)?;
-                    py_pack_param(&circuit, qpy_data, ValueEndian::LittleForV17AndBelow)
+        Parameters::Blocks(blocks) => {
+            qpy_data
+                .caller
+                .attach("control flow blocks", |py| -> Result<_, QpyError> {
+                    blocks
+                        .iter()
+                        .map(|block: &CircuitData| -> Result<_, QpyError> {
+                            // we explicitly name the block "unnamed" because otherwise it will be assigned a serial number name (e.g. "circuit-45")
+                            // which would result in inconsistent results, e.g. when packing the same circuit twice on the same run
+                            let py_block: PyCircuitData = block.clone().into();
+                            let circuit = py_block.into_py_quantum_circuit(py)?;
+                            py_pack_param(&circuit, qpy_data, ValueEndian::LittleForV17AndBelow)
+                        })
+                        .collect::<Result<_, QpyError>>()
                 })
-                .collect::<Result<_, QpyError>>()
-        }),
+        }
     }
 }
 
@@ -350,17 +354,21 @@ fn pack_pauli_product_measurement(
     // since we won't recreate this gate via python, it's not important to verify the python name is identical to the one we use here
     // so we simply hard-code it instead of going through python
     let gate_class_name = String::from(PAULI_PRODUCT_MEASUREMENT_GATE_CLASS_NAME);
-    let params = Python::attach(|py| -> Result<_, QpyError> {
-        let z_array = ppm.z.to_pyarray(py);
-        let x_array = ppm.x.to_pyarray(py);
-        // Pauli phase: 0 means +1, 2 means -1 (i.e. neg)
-        let phase: i64 = if ppm.neg { 2 } else { 0 };
-        Ok(vec![
-            py_pack_param(&z_array, qpy_data, ValueEndian::Big)?,
-            py_pack_param(&x_array, qpy_data, ValueEndian::Big)?,
-            pack_generic_value(&GenericValue::Int64(phase), qpy_data)?,
-        ])
-    })?;
+    let params = qpy_data.caller.attach(
+        "Pauli product measurement parameters",
+        |py| -> Result<_, QpyError> {
+            // TODO: get rid of the Python dependency here, we can serialize the arrays directly
+            let z_array = ppm.z.to_pyarray(py);
+            let x_array = ppm.x.to_pyarray(py);
+            // Pauli phase: 0 means +1, 2 means -1 (i.e. neg)
+            let phase: i64 = if ppm.neg { 2 } else { 0 };
+            Ok(vec![
+                py_pack_param(&z_array, qpy_data, ValueEndian::Big)?,
+                py_pack_param(&x_array, qpy_data, ValueEndian::Big)?,
+                pack_generic_value(&GenericValue::Int64(phase), qpy_data)?,
+            ])
+        },
+    )?;
     Ok(formats::CircuitInstructionV2Pack {
         num_qargs: instruction.op.num_qubits(),
         num_cargs: instruction.op.num_clbits(),
@@ -384,15 +392,19 @@ fn pack_pauli_product_rotation(
     // since we won't recreate this gate via python, it's not important to verify the python name is identical to the one we use here
     // so we simply hard-code it instead of going through python
     let gate_class_name = String::from(PAULI_PRODUCT_ROTATION_GATE_CLASS_NAME);
-    let params = Python::attach(|py| -> Result<_, QpyError> {
-        let z_array = rotation.z.to_pyarray(py);
-        let x_array = rotation.x.to_pyarray(py);
-        Ok(vec![
-            py_pack_param(&z_array, qpy_data, ValueEndian::Big)?,
-            py_pack_param(&x_array, qpy_data, ValueEndian::Big)?,
-            pack_param_obj(&rotation.angle, qpy_data, ValueEndian::LittleForV17AndBelow)?,
-        ])
-    })?;
+    let params = qpy_data.caller.attach(
+        "Pauli-product rotation parameters",
+        |py| -> Result<_, QpyError> {
+            // TODO: get rid of the Python dependency here, we can serialize the arrays directly
+            let z_array = rotation.z.to_pyarray(py);
+            let x_array = rotation.x.to_pyarray(py);
+            Ok(vec![
+                py_pack_param(&z_array, qpy_data, ValueEndian::Big)?,
+                py_pack_param(&x_array, qpy_data, ValueEndian::Big)?,
+                pack_param_obj(&rotation.angle, qpy_data, ValueEndian::LittleForV17AndBelow)?,
+            ])
+        },
+    )?;
     Ok(formats::CircuitInstructionV2Pack {
         num_qargs: instruction.op.num_qubits(),
         num_cargs: 0,
@@ -589,14 +601,18 @@ fn pack_unitary_gate(
 
     // until we change the QPY version or verify we get the exact same result,
     // we translate the matrix to numpy and then serialize it like python does
-    let params = Python::attach(|py| -> Result<_, QpyError> {
-        let out_array = matrix.to_pyarray(py);
-        Ok(vec![py_pack_param(
-            &out_array,
-            qpy_data,
-            ValueEndian::LittleForV17AndBelow,
-        )?])
-    })?;
+    // TODO: get rid of the Python dependency here, we can serialize the array directly
+    let params =
+        qpy_data
+            .caller
+            .attach("unitary gate parameters", |py| -> Result<_, QpyError> {
+                let out_array = matrix.to_pyarray(py);
+                Ok(vec![py_pack_param(
+                    &out_array,
+                    qpy_data,
+                    ValueEndian::LittleForV17AndBelow,
+                )?])
+            })?;
     // since we won't recreate this gate via python, it's not important to verify the python name is identical to the one we use here
     // so we simply hard-code it instead of going through python
     let gate_class_name = String::from(UNITARY_GATE_CLASS_NAME);
@@ -622,35 +638,48 @@ fn pack_py_instruction(
 ) -> Result<formats::CircuitInstructionV2Pack, QpyError> {
     let params = match py_inst.kind {
         PyOpKind::Gate | PyOpKind::Instruction => pack_instruction_params(instruction, qpy_data)?,
-        PyOpKind::Operation => Python::attach(|py| -> Result<_, QpyError> {
-            let py_op_object = py_inst.ob.bind(py);
-            if py_op_object.is_instance(imports::CLIFFORD.get_bound(py))? {
-                let tableau = py_op_object.getattr("tableau")?;
-                Ok(vec![py_pack_param(
-                    &tableau,
-                    qpy_data,
-                    ValueEndian::LittleForV17AndBelow,
-                )?])
-            } else if py_op_object.is_instance(imports::ANNOTATED_OPERATION.get_bound(py))? {
-                let modifiers = py_op_object.getattr("modifiers")?;
-                modifiers
-                    .try_iter()?
-                    .map(|modifier| {
-                        py_pack_param(&modifier?, qpy_data, ValueEndian::LittleForV17AndBelow)
-                    })
-                    .collect::<Result<_, QpyError>>()
-            } else {
-                pack_instruction_params(instruction, qpy_data)
-            }
-        })?,
+        PyOpKind::Operation => {
+            qpy_data
+                .caller
+                .attach("Python-defined operations", |py| -> Result<_, QpyError> {
+                    let py_op_object = py_inst.ob.bind(py);
+                    if py_op_object.is_instance(imports::CLIFFORD.get_bound(py))? {
+                        let tableau = py_op_object.getattr("tableau")?;
+                        Ok(vec![py_pack_param(
+                            &tableau,
+                            qpy_data,
+                            ValueEndian::LittleForV17AndBelow,
+                        )?])
+                    } else if py_op_object
+                        .is_instance(imports::ANNOTATED_OPERATION.get_bound(py))?
+                    {
+                        let modifiers = py_op_object.getattr("modifiers")?;
+                        modifiers
+                            .try_iter()?
+                            .map(|modifier| {
+                                py_pack_param(
+                                    &modifier?,
+                                    qpy_data,
+                                    ValueEndian::LittleForV17AndBelow,
+                                )
+                            })
+                            .collect::<Result<_, QpyError>>()
+                    } else {
+                        pack_instruction_params(instruction, qpy_data)
+                    }
+                })?
+        }
     };
+
     Ok(formats::CircuitInstructionV2Pack {
         num_qargs: py_inst.num_qubits(),
         num_cargs: py_inst.num_clbits(),
         extras_key: 0,
         num_ctrl_qubits: py_inst.num_ctrl_qubits().unwrap_or(0),
         ctrl_state: py_inst.ctrl_state().unwrap_or(0),
-        gate_class_name: Python::attach(|py| py_inst.class_name(py))?,
+        gate_class_name: qpy_data.caller.attach("Python defined instruction", |py| {
+            py_inst.class_name(py).map_err(QpyError::from)
+        })?,
         label: Default::default(),
         condition: Default::default(),
         bit_data: Default::default(),
@@ -975,18 +1004,20 @@ fn pack_custom_instructions(
 ) -> Result<formats::CustomCircuitInstructionsPack, QpyError> {
     let mut custom_instructions: Vec<formats::CustomCircuitInstructionDefPack> = Vec::new();
     let mut instructions_to_pack: Vec<String> = custom_instructions_hash.keys().cloned().collect();
-    Python::attach(|py| -> Result<_, QpyError> {
-        while let Some(name) = instructions_to_pack.pop() {
-            custom_instructions.push(pack_custom_instruction(
-                py,
-                &name,
-                custom_instructions_hash,
-                &mut instructions_to_pack,
-                qpy_data,
-            )?);
-        }
-        Ok(())
-    })?;
+    qpy_data
+        .caller
+        .attach("custom instructions", |py| -> Result<_, QpyError> {
+            while let Some(name) = instructions_to_pack.pop() {
+                custom_instructions.push(pack_custom_instruction(
+                    py,
+                    &name,
+                    custom_instructions_hash,
+                    &mut instructions_to_pack,
+                    qpy_data,
+                )?);
+            }
+            Ok(())
+        })?;
     Ok(formats::CustomCircuitInstructionsPack {
         custom_instructions,
     })
